finalized moving some smaller stuff around
