+++ crates/utiles-core/src/lib.rs
@@ -13,6 +13,8 @@
 #![expect(clippy::similar_names)]
 
 #[doc(inline)]
+pub use crate::constants::*;
+#[doc(inline)]
 pub use crate::fns::*;
 #[doc(inline)]
 pub use crate::merge::*;
@@ -50,6 +52,7 @@ pub mod parsing;
 
 mod asserts;
 mod edges;
+mod from;
 mod macros;
 mod merge;
 mod parent;
