+++ crates/utiles-cover/src/tests/mod.rs
@@ -1,1 +1,2 @@
+#![expect(clippy::unwrap_used)]
 mod coverage;
